refactor(sync-client): move vector-clock out of desktop as the first seam
Spec 001-mobile-app T018, first slice only.

Proves the extraction mechanism end to end: the module and its suite move to
packages/sync-client, 26 call sites switch to @memry/sync-client/vector-clock,
and the package joins the root typecheck and test filters. Nothing else moves —
the remaining files are blocked on the T015 vault-file-I/O seam decision.

Also fixes a semicolon inside a trailing comment in 0006_client_gate.sql: the
sync-harness migration runner splits on ';' after dropping only full-line
comments, so it tore the CREATE TABLE in two and every harness-backed desktop
integration test failed with D1_ERROR: incomplete input. Recorded in the
migrations README so the next migration does not repeat it.
