+++ apps/desktop/src/main/sync/item-handlers/canvas-handler.ts
@@ -9,7 +9,7 @@ import { CanvasSyncPayloadSchema, type CanvasSyncPayload } from '@memry/contract
 import { CanvasChannels } from '@memry/contracts/ipc-channels'
 import type { VectorClock } from '@memry/contracts/sync-api'
 import type { SyncQueueManager } from '../queue'
-import { increment } from '../vector-clock'
+import { increment } from '@memry/sync-client/vector-clock'
 import { createLogger } from '../../lib/logger'
 import { generateId } from '../../lib/id'
 import {
