+++ apps/desktop/src/main/sync/item-handlers/note-handler-sync-helpers.ts
@@ -2,7 +2,7 @@ import fs from 'fs'
 import { and, isNull, sql } from 'drizzle-orm'
 import { noteMetadata } from '@memry/db-schema/data-schema'
 import type { SyncQueueManager } from '../queue'
-import { increment } from '../vector-clock'
+import { increment } from '@memry/sync-client/vector-clock'
 import { extractFolderFromPath } from '../note-sync'
 import { isBinaryFileType } from '@memry/shared/file-types'
 import { toAbsolutePath } from '../../vault/notes'
