+++ apps/desktop/src/main/sync/item-handlers/reminder-handler.ts
@@ -5,7 +5,7 @@ import { ReminderChannels } from '@memry/contracts/ipc-channels'
 import type { VectorClock } from '@memry/contracts/sync-api'
 import { utcNow } from '@memry/shared/utc'
 import type { SyncQueueManager } from '../queue'
-import { increment } from '../vector-clock'
+import { increment } from '@memry/sync-client/vector-clock'
 import { createLogger } from '../../lib/logger'
 import { toOutboundReminderPayload } from '../reminder-outbound'
 import { BaseItemHandler } from './base-handler'
