+++ apps/desktop/src/main/sync/item-handlers/task-activity-handler.ts
@@ -8,7 +8,7 @@ import { TaskActivityChannels } from '@memry/contracts/ipc-channels'
 import type { VectorClock } from '@memry/contracts/sync-api'
 import { utcNow } from '@memry/shared/utc'
 import type { SyncQueueManager } from '../queue'
-import { increment } from '../vector-clock'
+import { increment } from '@memry/sync-client/vector-clock'
 import { createLogger } from '../../lib/logger'
 import { isBeyondTaskActivityRetention } from '../task-activity-retention'
 import { BaseItemHandler } from './base-handler'
