+++ apps/desktop/src/main/sync/item-handlers/task-handler.ts
@@ -8,7 +8,7 @@ import { TasksChannels } from '@memry/contracts/ipc-channels'
 import type { VectorClock, FieldClocks } from '@memry/contracts/sync-api'
 import { utcNow } from '@memry/shared/utc'
 import type { SyncQueueManager } from '../queue'
-import { increment } from '../vector-clock'
+import { increment } from '@memry/sync-client/vector-clock'
 import { mergeTaskFields, initAllFieldClocks, TASK_SYNCABLE_FIELDS } from '../field-merge'
 import { createLogger } from '../../lib/logger'
 import { BaseItemHandler } from './base-handler'
