+++ apps/desktop/src/main/sync/item-handlers/types.ts
@@ -3,7 +3,7 @@ import type { BetterSQLite3Database } from 'drizzle-orm/better-sqlite3'
 import type * as dataSchema from '@memry/db-schema/data-schema'
 import type { VectorClock, SyncItemType } from '@memry/contracts/sync-api'
 import type { SyncQueueManager } from '../queue'
-import { compare, merge } from '../vector-clock'
+import { compare, merge } from '@memry/sync-client/vector-clock'
 
 export type DrizzleDb = BetterSQLite3Database<typeof dataSchema>
 export type EmitToWindows = (channel: string, data: unknown) => void
