+++ package.json
@@ -38,7 +38,7 @@
     "typecheck:landing": "pnpm --filter @memry/landing typecheck",
     "typecheck:sync-server": "pnpm --filter @memry/sync-server typecheck",
     "typecheck:emails": "pnpm --filter @memry/marketing-emails typecheck",
-    "test": "pnpm repair:links && node scripts/run-turbo.js run test --filter=@memry/contracts --filter=@memry/app-core --filter=@memry/cli --filter=@memry/desktop --filter=@memry/editor-schema --filter=@memry/sync-server --filter=@memry/landing",
+    "test": "pnpm repair:links && node scripts/run-turbo.js run test --filter=@memry/contracts --filter=@memry/sync-client --filter=@memry/app-core --filter=@memry/cli --filter=@memry/desktop --filter=@memry/editor-schema --filter=@memry/sync-server --filter=@memry/landing",
     "test:cli": "pnpm --filter @memry/app-core test && pnpm --filter @memry/cli test",
     "test:release": "node --test scripts/release-utils.test.mjs scripts/release-notes-utils.test.mjs scripts/reddit-release-utils.test.mjs scripts/pre-push-policy.test.mjs scripts/docs-impact.test.mjs scripts/package-scripts.test.mjs apps/desktop/scripts/build-packaged-app-utils.test.cjs apps/desktop/scripts/check-packaged-runtime-deps-utils.test.cjs",
     "test:desktop": "pnpm --filter @memry/desktop test",
