+++ packages/sync-client/package.json
@@ -7,10 +7,12 @@
   "exports": {
     ".": "./src/index.ts",
     "./adapters": "./src/adapters/index.ts",
-    "./adapters/conformance": "./src/adapters/__tests__/conformance.ts"
+    "./adapters/conformance": "./src/adapters/__tests__/conformance.ts",
+    "./vector-clock": "./src/vector-clock.ts"
   },
   "types": "./src/index.ts",
   "scripts": {
+    "test": "vitest run",
     "typecheck": "tsc --noEmit -p tsconfig.json"
   },
   "dependencies": {
