+++ specs/001-mobile-app/tasks.md
@@ -69,7 +69,7 @@ additive and precedes any mobile write exposure. Verification: quickstart §Phas
 - [x] T015 Seam inventory: scan `apps/desktop/src/main/sync/` (314 files) + `packages/app-core` (18 node-touching) + `packages/storage-vault` (1); map every hit to one of the 10 seams (e.g. `network.ts` → http-client/runtime, `crdt-pending-notes.ts` → crdt-persistence); zero unassigned — `packages/sync-client/docs/seam-inventory.md` _(`packages/sync-client/docs/seam-inventory.md` + regenerable `scan-seams.mjs`. 174 non-test files scanned, 69 platform-touching, all assigned. **Three findings**: (1) vault file I/O (20 files) fits none of the ten seams — recommend widening `VaultDirectory` into `VaultFileSystem` rather than an 11th seam; **needs owner sign-off and blocks T018–T021**; (2) 23 `*-sync.ts` files are platform-bound only by the `BetterSQLite3Database` **type** — a type widening, no seam; (3) `node:events`/`node:crypto`/`node:os` are mechanical substitutions, not seams)_
 - [x] T016 Scaffold `packages/sync-client` (package.json with raw `./src` exports matching workspace conventions, tsconfig, turbo wiring) — `packages/sync-client/package.json` _(`packages/sync-client` — raw `./src` exports matching workspace convention, `base.json` tsconfig with `types: []` so node globals cannot leak in, registered in the root `typecheck` filter allowlist; `pnpm typecheck` 18/18 green)_
 - [x] T017 Define the 10 adapter interfaces exactly per [contracts/platform-adapters.md](./contracts/platform-adapters.md) — platform-free types only — `packages/sync-client/src/adapters/*.ts` _(one file per seam under `src/adapters/`, plus `SyncPlatformAdapters` and `SYNC_ADAPTER_SEAMS`. Lifted verbatim from contracts/platform-adapters.md; platform-free types only — `Uint8Array` for bytes, Promises for effects)_
-- [ ] T018 Move platform-free sync engine files (item-handlers registry, outbox logic, vector clocks, protocol client) into `packages/sync-client/src/` as import-path-only diffs, one seam per commit
+- [ ] T018 Move platform-free sync engine files (item-handlers registry, outbox logic, vector clocks, protocol client) into `packages/sync-client/src/` as import-path-only diffs, one seam per commit _(**STARTED, not complete.** `vector-clock.ts` + its suite moved as the first seam, proving the mechanism end-to-end: workspace wiring, `@memry/sync-client/vector-clock` subpath, 26 import-path-only edits, boundary check green. The remaining moves are **blocked on the T015 finding** — vault file I/O has no seam, so ~20 files cannot be assigned until `VaultDirectory` is widened or an 11th seam is approved.)_
 - [ ] T019 Move CRDT merge/pending logic behind the `CrdtPersistence` + `CrdtProvider` seams — `packages/sync-client/src/crdt/`
 - [ ] T020 Desktop adapter implementations (electron/node imports allowed **only** here) — `apps/desktop/src/main/sync/adapters/*.ts`
 - [ ] T021 App-core split: move the 18 node-touching files behind seams or into desktop; pure domain stays importable by mobile — `packages/app-core/src/`
